+++ sdk/hybridconnectivity/arm-hybridconnectivity/samples-dev/publicCloudConnectorsListByResourceGroupSample.ts
@@ -1,0 +1,31 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import { HybridConnectivityClient } from "@azure/arm-hybridconnectivity";
+import { DefaultAzureCredential } from "@azure/identity";
+
+/**
+ * This sample demonstrates how to list PublicCloudConnector resources by resource group
+ *
+ * @summary list PublicCloudConnector resources by resource group
+ * x-ms-original-file: 2024-12-01/PublicCloudConnectors_ListByResourceGroup.json
+ */
+async function publicCloudConnectorsListByResourceGroup() {
+  const credential = new DefaultAzureCredential();
+  const subscriptionId = "5ACC4579-DB34-4C2F-8F8C-25061168F342";
+  const client = new HybridConnectivityClient(credential, subscriptionId);
+  const resArray = new Array();
+  for await (let item of client.publicCloudConnectors.listByResourceGroup(
+    "rgpublicCloud",
+  )) {
+    resArray.push(item);
+  }
+
+  console.log(resArray);
+}
+
+async function main() {
+  publicCloudConnectorsListByResourceGroup();
+}
+
+main().catch(console.error);